requirements.txt
ag2==0.7.1
annotated-types==0.7.0
anyio==4.8.0
asyncer==0.0.8
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
fast-depends==2.4.12
FLAML==2.3.3
fonttools==4.55.4
frozendict==2.4.6
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
kiwisolver==1.4.8
lxml==5.3.0
matplotlib==3.10.0
multitasking==0.0.11
numpy==1.26.4
openai==1.59.9
packaging==24.2
pandas==2.2.3
peewee==3.17.8
pillow==11.1.0
platformdirs==4.3.6
pyautogen==0.7.1
pydantic==2.10.5
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
regex==2024.11.6
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
termcolor==2.5.0
tiktoken==0.8.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
webencodings==0.5.1
websockets==14.2
yfinance==0.2.52
